legalsolver/requirements.txt
flask==2.0.1
werkzeug==2.0.1
nltk==3.6.2
scikit-learn==0.24.2
torch==1.9.0
transformers==4.8.2
opencv-python==4.5.3.56
pytesseract==0.3.8
PyPDF2==1.26.0
python-docx==0.8.11
google-auth==2.6.0
google-auth-oauthlib==0.5.1
google-auth-httplib2==0.1.0
google-api-python-client==2.39.0
